refactor: move the status command to internal/cli — the read-command pilot (#529 slice 3)
Third slice of #529: the first source-agnostic READ COMMAND moves into
internal/cli, validating the command lift-and-shift pattern on the cheapest case
(status, ~114 LOC) before scaling to query/recover/reconstruct/shim.

- internal/cli/status.go: statusCmd + runStatus + flags moved verbatim from
  cmd/bintrail (package main -> cli). Its init() keeps the flags + required mark
  + BindCommandEnv, but no longer calls rootCmd.AddCommand.
- internal/cli/commands.go: new AddReadCommands(root) — the registration entry
  point each binary's main calls (today: status; grows as more commands move).
- cmd/bintrail/main.go: calls cli.AddReadCommands(rootCmd) in init().
- internal/cli/status_test.go: moved; TestStatusCmd_registered adapted to assert
  AddReadCommands(root) registers status (was a package-global rootCmd check).

No behavior change. Validated end-to-end: `go test -tags integration -run
TestEndToEnd` (which runs the real `bintrail status` binary) passes, plus unit +
vet + the cmd/bintrail uifree guard. Part of #529.
